Move DI container codes to more appropriate package and related changes to dockerfile
