+++ Dockerfile
@@ -7,7 +7,7 @@ COPY go.mod go.sum ./
 RUN go mod download
 
 COPY . .
-RUN CGO_ENABLED=0 GOOS=linux go build -o exchange-service ./cmd/web-service/
+RUN CGO_ENABLED=0 GOOS=linux go build -o exchange-service ./cmd/application/
 
 
 FROM gcr.io/distroless/static:nonroot
