+++ cmd/application/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"exchange-service/internal/app"
+)
+
+func main() {
+	app.NewApplication().Run()
+}
