+++ internal/app/application.go
@@ -1,11 +1,7 @@
-package main
+package app
 
 import (
-	"go.uber.org/fx"
-	"go.uber.org/fx/fxevent"
-
 	"exchange-service/internal/api"
-	"exchange-service/internal/app"
 	"exchange-service/internal/configuration"
 	"exchange-service/internal/database"
 	"exchange-service/internal/discovery"
@@ -16,15 +12,29 @@ import (
 	"exchange-service/internal/service/currency"
 	"exchange-service/internal/service/exchange"
 	"exchange-service/internal/util"
+
+	"go.uber.org/fx"
+	"go.uber.org/fx/fxevent"
 )
 
-func main() {
+type Application interface {
+	Run()
+}
+
+func NewApplication() Application {
+	return FxContainer{}
+}
+
+type FxContainer struct {
+}
+
+func (FxContainer) Run() {
 	fx.New(
 		fx.Provide(configuration.NewConfig),
-		fx.Provide(app.NewLogger),
-		fx.Provide(app.NewFxLogger),
-		fx.Provide(app.ProvideEcho),
-		fx.Provide(app.NewAppSetupManager),
+		fx.Provide(NewLogger),
+		fx.Provide(NewFxLogger),
+		fx.Provide(ProvideEcho),
+		fx.Provide(NewAppSetupManager),
 		fx.Provide(discovery.NewServiceDiscovery),
 		fx.Provide(util.NewValidator),
 		fx.Provide(util.NewHttpClient),
@@ -51,13 +61,13 @@ func main() {
 		asHandler(api.NewPriceHandler),
 		asHandler(api.NewOrderHandler),
 		fx.Provide(fx.Annotate(
-			app.NewRestApp,
+			NewRestApp,
 			fx.ParamTags(`group:"handlers"`),
 		)),
-		fx.WithLogger(func(log app.FxLogger) fxevent.Logger {
+		fx.WithLogger(func(log FxLogger) fxevent.Logger {
 			return &log
 		}),
-		fx.Invoke(app.ManageLifeCycle),
+		fx.Invoke(ManageLifeCycle),
 	).Run()
 }
